+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and this 
 
 ## 0.2.0
 
+- More information for blocks
 - Show help menu by pressing `?`
 - Options to add unique emoji for each course to remember & find needed courses quicker
 - Option to hide courses #2